chore: adjust naming variable and conditional method

+++ app/Services/Backend/ImageService.php
@@ -12,11 +12,13 @@ class ImageService
      */
     public function findImage(object $data, string $section) 
     {
-        $path = 'img/uploads/' . $section . '/' . $data->image;
-        
-        if (File::exists($path)) {
-            File::delete($path);
-        } 
+        if(!is_null($data->image)) {
+            $path = 'img/uploads/' . $section . '/' . $data->image;
+            
+            if (File::exists($path)) {
+                File::delete($path);
+            } 
+        }
     }
 
     /**
